Fix rebuild

+++ utils/rebuild.sh
@@ -7,12 +7,16 @@ cd  ~/foodList
 git checkout main
 git pull origin main
 
-sudo docker rm -f $(docker ps -a -q)
+# Check if there are running containers
+if [ "$(docker ps -q)" != "" ]; then
+    # Stop all running containers
+    sudo docker rm -f $(docker ps -a -q)
+fi
 
 # Check if there are any Docker images
 if [ "$(docker images -q)" != "" ]; then
     # Remove all Docker images
-    sudo docker rmi -f $(docker images -q)
+    sudo docker rmi $(docker images -q)
 fi
 
 docker image prune -f
